feat: Share Connector Instance with Sinks (#184)
Small changes to enable sharing the connector instance, which was
recently added to the sdk. Used to show a few hundred connections in
postgres when syncing Salesforce data, now only seeing one.

Also removed two methods that were deprecated (and unused) or redundant
after updates to the SDK.

I haven't tested how this impacts speed and I don't know how to create a
test for this. Happy to do so if pointed in the right direction.

---------

Co-authored-by: Derek Visch <visch@users.noreply.github.com>