feat: Optionally store content marked as `"base16"` encoded into `bytea` columns (#287)
Motivation: more efficient `bytea` storage of data marked with
`"contentEncoding": "base16"`

JSON Schema References:
- https://datatracker.ietf.org/doc/html/rfc4648#section-8
-
https://json-schema.org/draft/2020-12/draft-bhutton-json-schema-validation-00#rfc.section.8.3

Bytea is at least twice as efficient as string to store hex data:
```
select 
    octet_length('\x2BdfBd329984Cf0DC9027734681A16f542cF3bB4'::bytea) as "bytea", 
    octet_length('0x2BdfBd329984Cf0DC9027734681A16f542cF3bB4') as "string"
;
 bytea | string 
-------+--------
    20 |     42
```

It's probably a good idea to make this behaviour opt-in. I'm not sure
how to implement that since most of the type detection code is inside
static methods of `PostgresConnector`. I can't find a way to inject the
target config without making a big change.

---------

Co-authored-by: Edgar Ramírez Mondragón <16805946+edgarrmondragon@users.noreply.github.com>