feat: Support string and int arrays (#330)
Like the title said, this adds support for
- `ARRAY(BIGINT())`
- `ARRAY(TEXT())`

Technically this is a breaking change. Where previously all array types
were `ARRAY(JSONB())`, now there's more nuance. Unlikely that many
people rely on this behaviour, but we had to write a migration. Should
we increment the major version (or at least the minor) to signal this
change? Or introduce a feature switch?

---------

Co-authored-by: Ruben Vereecken <rubenvereecken@gmail.com>
Co-authored-by: Edgar Ramírez Mondragón <16805946+edgarrmondragon@users.noreply.github.com>