feat: Configurable null character sanitization (#434)
Null characters are currently passed as-is to Postgres despite being
unsupported.

If it is encountered, it causes the sink to fail as noted here:
https://github.com/MeltanoLabs/target-postgres/issues/60 with an error
like `ValueError: A string literal cannot contain NUL (0x00)
characters.`

This PR introduces a new option called `sanitize_null_text_characters`
which enables sanitization of these characters.

---------

Co-authored-by: Edgar Ramírez Mondragón <16805946+edgarrmondragon@users.noreply.github.com>
Co-authored-by: Edgar Ramírez-Mondragón <edgarrm358@gmail.com>